+++ android/app/src/gmsStubs/java/com/google/android/gms/common/api/Scope.java
@@ -1,7 +1,0 @@
-package com.google.android.gms.common.api;
-
-public class Scope {
-    public Scope(String myScope) {
-
-    }
-}