+++ android/app/src/main/java/com/foobnix/model/AppSP.java
@@ -54,7 +54,6 @@ public class AppSP {
     public boolean isCrop = false;
     public boolean isCropSymetry = false;
     public boolean isSmartReflow = false;
-    public boolean isEnableSync;
     public String syncRootID;
 
     public String currentProfile = "";