+++ android/app/src/main/java/com/foobnix/ui2/fragment/PrefFragment2.java
@@ -64,7 +64,6 @@ import com.foobnix.android.utils.ResultResponse2;
 import com.foobnix.android.utils.TxtUtils;
 import com.foobnix.android.utils.Views;
 import com.foobnix.dao2.FileMeta;
-import com.foobnix.drive.GFile;
 import com.foobnix.ext.CacheZipUtils;
 import com.foobnix.model.AppProfile;
 import com.foobnix.model.AppSP;
@@ -102,8 +101,6 @@ import com.foobnix.pdf.info.widget.ShareDialog;
 import com.foobnix.pdf.info.wrapper.DocumentController;
 import com.foobnix.pdf.info.wrapper.PasswordState;
 import com.foobnix.pdf.info.wrapper.UITab;
-import com.foobnix.pdf.search.activity.msg.GDriveSycnEvent;
-import com.foobnix.pdf.search.activity.msg.MessageSync;
 import com.foobnix.sys.TempHolder;
 import com.foobnix.ui2.AdsFragmentActivity;
 import com.foobnix.ui2.AppDB;
@@ -112,7 +109,6 @@ import com.foobnix.ui2.MainTabs2;
 import com.foobnix.ui2.MyContextWrapper;
 import com.foobnix.webdav.WebDavSyncDialog;
 import com.foobnix.work.SearchAllBooksWorker;
-import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.jmedeisis.draglinearlayout.DragLinearLayout;
 
 import org.greenrobot.eventbus.Subscribe;
@@ -138,8 +134,6 @@ public class PrefFragment2 extends UIFragment {
     private static final String WWW_WIKI_SITE = "https://github.com/380121850/LibreraReader";
     View section1, section2, section4, section8, section9, panelRecent, overlay,
             statusBarHack;
-    TextView singIn, syncInfo, syncInfo2, syncHeader;
-    CheckBox isEnableSync;
     private TextView curBrightness, themeColor, profileLetter;
     private CheckBox isRememberDictionary;
     private TextView nextKeys;
@@ -211,134 +205,11 @@ public class PrefFragment2 extends UIFragment {
 
     }
 
-    @Subscribe public void updateSyncInfo(GDriveSycnEvent event) {
-        String gdriveInfo = GFile.getDisplayInfo(getActivity());
-
-        if (TxtUtils.isEmpty(gdriveInfo)) {
-            AppSP.get().isEnableSync = false;
-            syncInfo.setVisibility(View.GONE);
-            singIn.setText(R.string.sign_in);
-            TxtUtils.underlineTextView(singIn);
-            singIn.setOnClickListener(new OnClickListener() {
-                @Override public void onClick(View v) {
-                    GFile.init(getActivity());
-                    updateSyncInfo(null);
-                }
-            });
-        } else {
-            syncInfo.setVisibility(View.VISIBLE);
-            syncInfo.setText(gdriveInfo);
-            singIn.setText(R.string.sign_out);
-            TxtUtils.underlineTextView(singIn);
-
-            singIn.setOnClickListener(new OnClickListener() {
-                @Override public void onClick(View v) {
-                    AppSP.get().isEnableSync = false;
-                    AppSP.get().syncRootID = "";
-                    AppSP.get().syncTime = 0;
-                    GFile.logout(getActivity());
-                    updateSyncInfo(null);
-                }
-            });
-        }
-
-        isEnableSync.setChecked(AppSP.get().isEnableSync);
-        onSync(null);
-
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN) public void onSync(MessageSync msg) {
-        if (AppSP.get().syncTime > 0) {
-
-            final Date date = new Date(AppSP.get().syncTime);
-            String format = "";
-            if (DateUtils.isToday(AppSP.get().syncTime)) {
-                format = getString(R.string.today) + " " + DateFormat.getTimeInstance()
-                                                                     .format(date);
-            } else {
-                format = DateFormat.getDateTimeInstance()
-                                   .format(date);
-            }
-
-            String status = AppSP.get().syncTimeStatus == MessageSync.STATE_SUCCESS ? getString(R.string.success) :
-                    getString(R.string.fail);
-            if (AppSP.get().syncTimeStatus == MessageSync.STATE_VISIBLE) {
-                status = "...";
-            }
-
-            syncInfo2.setText(format + " - " + status);
-            syncInfo2.setVisibility(View.VISIBLE);
-        } else {
-            syncInfo2.setText("");
-            syncInfo2.setVisibility(View.GONE);
-            syncHeader.setText(R.string.sync_google_drive);
-
-        }
-    }
-
     @Override public View onCreateView(final LayoutInflater inflater, final ViewGroup container,
                                        final Bundle savedInstanceState) {
         inflate = inflater.inflate(R.layout.fragment_preferences, container, false);
 
-        singIn = inflate.findViewById(R.id.signIn);
-        syncInfo = inflate.findViewById(R.id.syncInfo);
-        syncInfo2 = inflate.findViewById(R.id.syncInfo2);
-        syncHeader = inflate.findViewById(R.id.syncHeader);
-        onSync(null);
-        syncHeader.setOnClickListener((in) -> Dialogs.showSyncLOGDialog(getActivity()));
-
-        isEnableSync = inflate.findViewById(R.id.isEnableSync);
-        // Google Drive sync is removed from all builds (Drive SDK stripped);
-        // hide the sync section so no entry point reaches the inert stubs.
-        singIn.setVisibility(View.GONE);
-        syncInfo.setVisibility(View.GONE);
-        syncInfo2.setVisibility(View.GONE);
-        syncHeader.setVisibility(View.GONE);
-        isEnableSync.setVisibility(View.GONE);
-        isEnableSync.setChecked(AppSP.get().isEnableSync);
-        isEnableSync.setOnCheckedChangeListener((buttonView, isChecked) -> {
-            AppSP.get().isEnableSync = isChecked;
-            if (isChecked && getActivity() != null) {
-                if (GoogleSignIn.getLastSignedInAccount(getActivity()) == null) {
-                    GFile.init(getActivity());
-                } else {
-                    GFile.runSyncService(getActivity());
-                }
-            }
-        });
-
-        View syncSettings = inflate.findViewById(R.id.isEnableSyncSettings);
-        syncSettings.setVisibility(View.GONE);
-        syncSettings.setOnClickListener(v -> {
-                   final CheckBox isSyncPullToRefresh = new CheckBox(getActivity());
-                   isSyncPullToRefresh.setText(R.string.pull_to_start_sync);
-                   isSyncPullToRefresh.setChecked(BookCSS.get().isSyncPullToRefresh);
-                   isSyncPullToRefresh.setOnCheckedChangeListener(
-                           (buttonView, isChecked) -> BookCSS.get().isSyncPullToRefresh = isChecked);
-
-                   final CheckBox isSyncWifiOnly = new CheckBox(getActivity());
-                   isSyncWifiOnly.setText(R.string.wifi_sync_only);
-                   isSyncWifiOnly.setChecked(BookCSS.get().isSyncWifiOnly);
-                   isSyncWifiOnly.setOnCheckedChangeListener(
-                           (buttonView, isChecked) -> BookCSS.get().isSyncWifiOnly = isChecked);
-
-                   final CheckBox isShowSyncWheel = new CheckBox(getActivity());
-                   isShowSyncWheel.setText(getString(R.string.animate_sync_progress));
-                   isShowSyncWheel.setChecked(BookCSS.get().isSyncAnimation);
-                   isShowSyncWheel.setOnCheckedChangeListener(
-                           (buttonView, isChecked) -> BookCSS.get().isSyncAnimation = isChecked);
-
-                   AlertDialogs.showViewDialog(getActivity(), null, isSyncPullToRefresh, isSyncWifiOnly,
-                           isShowSyncWheel);
-               });
-
-        updateSyncInfo(null);
-
         section8 = inflate.findViewById(R.id.section8);
-
-        inflate.findViewById(R.id.sectionSync)
-               .setVisibility(AppsConfig.IS_FDROID ? View.GONE : View.VISIBLE);
-        //inflate.findViewById(R.id.sectionSync).setVisibility(View.GONE);//TODO GDIVE need to fix
 
         section9 = inflate.findViewById(R.id.section9);
         panelRecent = inflate.findViewById(R.id.panelRecent);