+++ android/app/src/main/java/com/google/api/client/extensions/android/http/AndroidHttp.java
@@ -1,7 +1,0 @@
-package com.google.api.client.extensions.android.http;
-
-public class AndroidHttp {
-    public static Object newCompatibleTransport() {
-        return null;
-    }
-}