+++ android/app/src/main/java/com/google/api/client/googleapis/extensions/android/gms/auth/UserRecoverableAuthIOException.java
@@ -1,4 +1,0 @@
-package com.google.api.client.googleapis.extensions.android.gms.auth;
-
-public class UserRecoverableAuthIOException extends Exception {
-}