+++ android/app/src/main/java/com/google/api/client/http/ByteArrayContent.java
@@ -1,7 +1,0 @@
-package com.google.api.client.http;
-
-public class ByteArrayContent {
-    public static ByteArrayContent fromString(String s, String content) {
-        return null;
-    }
-}