+++ android/app/src/main/java/com/google/api/client/http/FileContent.java
@@ -1,8 +1,0 @@
-package com.google.api.client.http;
-
-import java.io.File;
-
-public class FileContent {
-    public FileContent(String mimeType, File inFile) {
-    }
-}