+++ android/app/src/main/java/com/google/api/client/http/javanet/NetHttpTransport.java
@@ -1,4 +1,0 @@
-package com.google.api.client.http.javanet;
-
-public class NetHttpTransport {
-}