+++ android/app/src/main/java/com/google/api/client/json/gson/GsonFactory.java
@@ -1,4 +1,0 @@
-package com.google.api.client.json.gson;
-
-public class GsonFactory {
-}